pyproject.toml
[project]
name = "kiro-mkdocs"
version = "0.1.0"
description = "Kiro × MkDocs Docs-as-Code 検証プロジェクト"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "mkdocs>=1.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
